Extract CliHelper
