+++ lib/tapioca/internal.rb
@@ -8,6 +8,7 @@ require "tapioca/sorbet_ext/fixed_hash_patch"
 require "tapioca/generic_type_registry"
 require "tapioca/config"
 require "tapioca/config_builder"
+require "tapioca/helpers/cli_helper"
 require "tapioca/generators"
 require "tapioca/cli"
 require "tapioca/gemfile"
